+++ Modules/Accounts/Components/Points/PointLoadFailedMessage.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Threading;
 using System.Threading.Tasks;
 using Unity.AI.Toolkit.Accounts.Services;
 using Unity.AI.Toolkit.Accounts.Services.Core;
@@ -8,17 +7,17 @@ using UnityEngine.UIElements;
 namespace Unity.AI.Toolkit.Accounts.Components
 {
     [UxmlElement]
-    partial class AccountLoadDelayedBanner : BasicBannerContent
+    partial class PointLoadFailedMessage : BasicBannerContent
     {
         Button m_Retry;
         DropdownLoading m_DropdownLoading;
         Image m_WarningIcon;
         RichLabel m_RichLabel;
 
-        public AccountLoadDelayedBanner()
+        public PointLoadFailedMessage()
         {
-            var message = $"Unable to load account information from {AccountApi.selectedEnvironment}";
-            var loadingMessage = "Loading account information";
+            var message = $"Unable to load point balance from {AccountApi.selectedEnvironment}";
+            var loadingMessage = "Loading points information";
 
             Init();
             content.AddToClassList("banner-content");
@@ -35,7 +34,7 @@ namespace Unity.AI.Toolkit.Accounts.Components
             content.Clear();
             content.Add(m_DropdownLoading);
 
-            await EditorTask.RunOnMainThread(Account.settings.RefreshSettings);
+            await EditorTask.RunOnMainThread(Account.pointsBalance.RefreshPointsBalance);
 
             content.Clear();
             content.Add(m_WarningIcon);