+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.16] - 2025-07-16
+
+### Fixed
+
+- Better handling of settings and points fetching errors.
+- Cleaner points and settings initialization when Editor Application is not in Focus.
+
 ## [1.0.0-pre.15] - 2025-06-27
 
 ### Changed